+++ object-storage-core/src/main/java/io/micronaut/objectstorage/response/UploadResponse.java
@@ -36,6 +36,8 @@ public interface UploadResponse<R> {
      * @param eTag the entity tag of the object stored.
      * @param nativeResponse upload response object.
      * @param <R> Cloud vendor-specific upload response
+     *
+     * @return an upload response instance.
      */
     @NonNull
     static <R> UploadResponse<R> of(@NonNull String key, @NonNull String eTag, @NonNull R nativeResponse) {
